Xcode: Added EE configurations & xcconfig files. Conditionalized SQLite.
* "Debug-EE" and "Release-EE" configurations are identical to "Debug"
  and "Release" except that they use the "Project-Debug-EE.xcconfig"
  and "Project-Release-EE.xcconfig" configuration files.
* Those config files override the preprocessor symbol settings to define
  COUCHBASE_ENTERPRISE.
* The SQLite target compiles a new source file SQLiteChooser.c instead of
  sqlite3.c. This new file uses "#ifdef COUCHBASE_ENTERPRISE" to
  conditionally include either sqlite3.c or the SEE equivalent.

+++ LiteCore/Storage/SQLiteDataFile.cc
@@ -298,7 +298,7 @@ namespace litecore {
             slice key = options().encryptionKey;
             if(key.buf == nullptr || key.size != 32)
                 error::_throw(error::InvalidParameter);
-            sqlite3_key_v2(_sqlDb->getHandle(), nullptr, key.buf, key.size);
+            sqlite3_key_v2(_sqlDb->getHandle(), nullptr, key.buf, (int)key.size);
 #else
             error::_throw(error::UnsupportedOperation);
 #endif
@@ -341,7 +341,7 @@ namespace litecore {
         if(alg == kNoEncryption) {
             rekeyResult = sqlite3_rekey_v2(_sqlDb->getHandle(), nullptr, nullptr, 0);
         } else {
-            rekeyResult = sqlite3_rekey_v2(_sqlDb->getHandle(), nullptr, newKey.buf, newKey.size);
+            rekeyResult = sqlite3_rekey_v2(_sqlDb->getHandle(), nullptr, newKey.buf, (int)newKey.size);
         }
         
         if(rekeyResult != SQLITE_OK) {
